renamed index.html to avoid a 301 and enable client side page caching
The Go file server by default redirects index.html which is a minor latency hit.  The caching issue is more interesting.  As of this writing all pages are assumeed to be static, and with a fixed url like /i/1234 they are (assuming no pics deleted).  I plan on enabling the cache for all index urls except for / to ensure hard urls are cacheable and the front page is fresh.  Just the next button will change.