Add uniqueness checks for images, and refactor tests to be simpler.
This is motivated by the expected upload process of automated scripts.  They may try and upload the same file multiple times, non maliciously.  This prevents that from happening.  No effort is made to prevent malicious uploads, or even very similar uploads.  A future idea is to merge the tags in the event of a duplicate image