be/imaging: fix bad thumbnail when gif frames are different sizes
When an image has frames of different sizes and offsets, pixur
used the last frame to pick the size, but cropped and resized
the first frame.  This changes the flow to always reset the iterator
for the picture after modifying it, so it is always ready for
the next call.

Also, added a test (which took forever to figure out how to repro).