be/schema/tables: qualify indexes with trailing id
This is needed when scanning the set with the possibility of
duplicate index values.  In order to paginate using just the
id of the pic, the scan needs to do an initial look up of the
pic, find the index order (or sched or sort order), and begin
the scan from *that* index.  The problem is that there may be
many pics with an identical index, so it may not be possible
to quickly find the starting point with the original id.

Instead, all indexes that are meant for pagination should have
as trailing unique identifier, which is used as a tie breaker
in the event of duplicates.  This allows finding pics to cost
1 initial lookup, and two scans to find out the next and previous
indexes for pagination.

Make them index still, rather than unique, to avoid insertion
cost.