be/schema/db: provide better error messages for postgres and make unique violations retryable
When importing an old database, it was possible for comments to have the
same timestamp.  When creating user events for this, they would all try
to allocate the same index for a given timestamp/user.  This results in
trying to insert an already existing row.  Instead, just treat it as
retryable.  In the event the tx could succeed, it will.  If it would not
have worked, this won't hide any bugs since the retry loop eventually
gives up.