be/schema/db: fix lock ordering deadlock in idalloc
ID alloc attempts to allocate IDs in a job separate from
the caller, to avoid blocking unnecessarily.  Previously,
a mutex would be acquired to see if there were any IDs and
if not, create a job to allocate one.  This causes a
deadlock because the SQL library in Go acquires a lock as well.

When the sql library does connection pooling, there can only be
at most ~20 active transactions at once.  If unlucky, each of
these connection can be in use, typically because a tx has been
started, but they are all waiting on the id alloc mutex. In this
case, all connections are held in the pool semaphore, and they
are all waiting to acquire the id alloc mutex.

Simulataneously, one thread wins the mutex race and attempts to
start a new transaction.  Because neither side will release their
lock, it's a deadlock.

To solve this, only briefly hold the lock to see if there are
any available IDs.  If not, drop the lock, start a connection,
and reacquire the lock.  Refill only if there are still no IDs
available.  This forces all threads to follow the same lock
acquisition pattern, and avoids the deadlock.

Additionally, this change also modifies AllocIDJob to use IDs
from the alloc.  Previously it would not, as the job could be
rolled back.  This is still true, but now it can consume the
ids from the IDAlloc itself, and simply not update the DB if
none are available.

Also, because transactions are rapidly rolled back with an
alloc grab size of 1, change it to 10.  1 was picked to ensure
correctness.  The impl is still correct, but I don't like seeing
all the suppressed retries from CockroachDB.